+++ python-code/src/amr_hub_abm/space/room.py
@@ -14,10 +14,8 @@ from amr_hub_abm.exceptions import InvalidRoomError, SimulationModeError
 from amr_hub_abm.space.location import Location
 
 if TYPE_CHECKING:
-    from matplotlib.axes import Axes
     from numpy.random import Generator
 
-    from amr_hub_abm.agent.agent import Agent
     from amr_hub_abm.space.content import Content
     from amr_hub_abm.space.door import Door
     from amr_hub_abm.space.wall import Wall
@@ -118,63 +116,6 @@ class Room:
             raise InvalidRoomError(msg)
 
         return polygon[0]
-
-    def plot(
-        self,
-        ax: Axes,
-        agents: list[Agent] | None = None,
-        *,
-        trajectory: bool = False,
-        **kwargs: dict,
-    ) -> None:
-        """Plot the room on a given matplotlib axis."""
-        if not self.walls:
-            msg = "Cannot plot room without walls."
-            raise SimulationModeError(msg)
-
-        for wall in self.walls:
-            wall.plot(ax, color="black")  # type: ignore  # noqa: PGH003
-
-        for door in self.doors:
-            x, y = door.line.xy
-            ax.plot(
-                x,
-                y,
-                color=kwargs.get("door_color", "brown"),
-                linewidth=kwargs.get("door_width", 2),
-            )
-
-        for content in self.contents:
-            ax.scatter(
-                content.position[0],
-                content.position[1],
-                marker=content.marker_type,
-                color=content.color,
-                s=content.marker_size,
-                label=f"{content.content_type.name} ({content.content_id})",
-            )
-            ax.text(
-                content.position[0] + 0.05,
-                content.position[1] - 0.15,
-                content.content_type.name.lower(),  # "bed", "chair", "workstation"
-                fontsize=6,
-                ha="left",
-                va="top",
-                color="gray",
-                alpha=0.7,
-            )
-
-        if agents is None:
-            return
-
-        for agent in agents:
-            if (
-                agent.location.building == self.building
-                and agent.location.floor == self.floor
-            ) and self.contains_point((agent.location.x, agent.location.y)):
-                agent.plot_agent(ax)
-                if trajectory and agent.agent_type.value == 2:  # HEALTHCARE_WORKER only
-                    agent.plot_trajectory(ax)
 
     def contains_point(self, point: tuple[float, float]) -> bool:
         """Check if a given point is inside the room."""
