+++ python-code/tests/test_room.py
@@ -1,7 +1,6 @@
 """Tests for the Room class in the AMR Hub ABM simulation."""
 
 from pathlib import Path
-from unittest.mock import patch
 
 import matplotlib.pyplot as plt
 import numpy as np
@@ -14,6 +13,7 @@ from amr_hub_abm.space.building import Building
 from amr_hub_abm.space.content import Content
 from amr_hub_abm.space.door import Door
 from amr_hub_abm.space.location import Location
+from amr_hub_abm.space.plotter import plot_agents_in_room
 from amr_hub_abm.space.room import Room
 from amr_hub_abm.space.wall import Wall
 
@@ -258,8 +258,8 @@ def test_invalid_plot_creation(
 
     fig = plt.figure()
     ax = fig.add_subplot(111)
-    with pytest.raises(SimulationModeError) as exc_info:
-        room.plot(ax=ax)
+    with pytest.raises(InvalidRoomError) as exc_info:
+        plot_agents_in_room(room, ax, agents=[])
     assert "Cannot plot room without walls." in str(exc_info.value)
 
 
@@ -365,7 +365,7 @@ def test_unconnected_walls() -> None:
 def test_plot_room(simple_room: Room) -> None:
     """Test plotting a room."""
     fig, ax = plt.subplots()
-    simple_room.plot(ax=ax)
+    plot_agents_in_room(simple_room, ax, agents=[])
     plt.close(fig)  # Close the plot to avoid displaying during tests
 
 
@@ -379,27 +379,8 @@ def test_plot_room_with_agent_inside(simple_room: Room) -> None:
         rooms=[simple_room],
         rng_generator=np.random.default_rng(),
     )
-    with patch.object(agent, "plot_agent") as mock_plot_agent:
-        simple_room.plot(ax=ax, agents=[agent])
-        mock_plot_agent.assert_called_once()
+    plot_agents_in_room(simple_room, ax, agents=[agent])
 
-    plt.close(fig)
-
-
-def test_plot_room_skips_agent_outside(simple_room: Room) -> None:
-    """Test plotting skips agents located outside the room."""
-    fig, ax = plt.subplots()
-    agent = Agent(
-        idx=2,
-        location=Location(6.0, 6.0, floor=1, building=simple_room.building),
-        heading_rad=0.0,
-        rooms=[simple_room],
-        rng_generator=np.random.default_rng(),
-    )
-
-    with patch.object(agent, "plot_agent") as mock_plot_agent:
-        simple_room.plot(ax=ax, agents=[agent])
-        mock_plot_agent.assert_not_called()
     plt.close(fig)
 
 
@@ -482,7 +463,7 @@ def test_room_plotting_with_doors(
     )
 
     fig, ax = plt.subplots()
-    room.plot(ax=ax)
+    plot_agents_in_room(room, ax, agents=[])
     if not Path("tests/output/").exists():
         Path("tests/output/").mkdir(parents=True, exist_ok=True)
     plt.savefig("tests/output/room_with_door_plot.png")
